Replace FeedbackUsuario with specific feedback classes
