+++ src/dao/UsuarioDAO.java
@@ -11,7 +11,6 @@ import java.sql.Statement;
 import java.sql.PreparedStatement;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
-import proyectobd.ParametrosGenerales.FeedbackUsuario;
 
 
 
@@ -21,8 +20,6 @@ import proyectobd.ParametrosGenerales.FeedbackUsuario;
  * @author admin
  */
 public class UsuarioDAO {
-    // Instanciamos FeedbackUsuario para mostrar las alertas
-    FeedbackUsuario fu=new FeedbackUsuario();
     
     // Método para obtener la lista de usuarios de la BD
     // Devolverá un array de UsuarioDTO
@@ -47,7 +44,7 @@ public class UsuarioDAO {
             }
             ConnBD.CerrarConexionBD();
         }catch(Exception ex){
-            fu.MostrarAlertas("Error: "+ex.toString());
+            System.err.println("Error: "+ex.toString());
         }
         return usuarios;
     }
@@ -76,7 +73,7 @@ public class UsuarioDAO {
             }
             ConnBD.CerrarConexionBD();
         }catch(Exception ex){
-            fu.MostrarAlertas("Error: "+ex.toString()+"\n"+sql);
+            System.err.println("Error: "+ex.toString()+"\n"+sql);
         }
         return usuarios;
     }
@@ -96,11 +93,11 @@ public class UsuarioDAO {
             if(rs.next()){
                 codigoInsertado=rs.getInt(1);
             }
-            fu.MostrarAlertas("Información del sistema", "Usuario registrado con ID: "+codigoInsertado);
+            System.out.println("Usuario registrado con ID: "+codigoInsertado);
             ConnBD.CerrarConexionBD();
             return codigoInsertado;
         }catch(Exception ex){
-            fu.MostrarAlertas("Error del sistema", ex.toString());
+            System.err.println("Error del sistema: "+ex.toString());
             return 0;
         }
     }
@@ -116,11 +113,11 @@ public class UsuarioDAO {
             ps.setString(4, usrdto.getContrasena());
             ps.setInt(5, usrdto.getId());
             int registrosActualizados=ps.executeUpdate();
-            fu.MostrarAlertas("Información del sistema", "Usuario actualizado. Registros modificados: "+registrosActualizados);
+            System.out.println("Usuario actualizado. Registros modificados: "+registrosActualizados);
             ConnBD.CerrarConexionBD();
             return registrosActualizados;
         }catch(Exception ex){
-            fu.MostrarAlertas("Error del sistema", ex.toString());
+            System.err.println("Error del sistema: "+ex.toString());
             return 0;
         }
     }
